Rename GamezPane to BoardPage
