+++ ui/src/BoardPage/PieceAttachmentsDialog.svelte
