+++ ui/src/BoardPage/PlayersBar.svelte
