+++ ui/src/BoardPage/WalAsset.svelte
