+++ ui/src/BoardPage/WalSpace.svelte
