--- ui/src/GamezPane/index.ts
@@ -1,1 +1,0 @@
-export { default as default } from './GamezPane.svelte';
